refactor: restructure invitation management pages
